Added support for building emulation in Windows.
It has not been tested but it should not break any linux build.

+++ src/asn1.h
@@ -27,17 +27,17 @@
 #endif
 
 extern int walk_tlv(const uint8_t *cdata,
-                    size_t cdata_len,
+                    uint16_t cdata_len,
                     uint8_t **p,
                     uint16_t *tag,
-                    size_t *tag_len,
+                    uint16_t *tag_len,
                     uint8_t **data);
-extern int format_tlv_len(size_t len, uint8_t *out);
+extern uint8_t format_tlv_len(uint16_t len, uint8_t *out);
 extern bool asn1_find_tag(const uint8_t *data,
-                          size_t data_len,
+                          uint16_t data_len,
                           uint16_t itag,
-                          size_t *tag_len,
+                          uint16_t *tag_len,
                           uint8_t **tag_data);
-extern size_t asn1_len_tag(uint16_t tag, size_t len);
+extern uint16_t asn1_len_tag(uint16_t tag, uint16_t len);
 
 #endif
